begin breaking up and reorganizing fig_helper and fig_metadata by moving some functions into figdata_transformer modules
