+++ app/fig_builder/plotly_scatter.py
@@ -3,9 +3,10 @@ import plotly.express as px
 import plotly.graph_objects as go
 from sklearn.manifold import TSNE
 
-import app.fig_builder.fig_helper as fh
 import app.fig_builder.fig_metadata as fm
-from app import utils
+import app.figdata_transformer.color_processor as cp
+import app.figdata_transformer.pandas_transformer as pt
+import app.figdata_transformer.speaker_topic_scatter_meta as stsm
 
 
 def build_episode_similarity_scatter(df: pd.DataFrame, seasons: list) -> go.Figure:
@@ -17,7 +18,7 @@ def build_episode_similarity_scatter(df: pd.DataFrame, seasons: list) -> go.Figu
     symbol_map = {'all': 'square', 'mlt': 'circle', 'focal': 'star'}
 
     # ad-hoc method of flattening topic metadata for hovertemplate display
-    df['flattened_topics'] = df['topics_universal_tfidf'].apply(fh.flatten_topics)
+    df['flattened_topics'] = df['topics_universal_tfidf'].apply(pt.flatten_topics)
 
     custom_data = ['title', 'season', 'episode', 'score', 'rank', 'focal_speakers', 'flattened_topics']
     
@@ -59,23 +60,23 @@ def build_all_series_episodes_scatter(df: pd.DataFrame, seasons: list, hilite: s
     x_high = x_max + buffer
 
     # ad-hoc method of flattening topic metadata for hovertemplate display
-    df['flattened_topics_tfidf'] = df.apply(lambda x: fh.flatten_topics(x['topics_universal_tfidf'], parent_only=True), axis=1)
-    df['flattened_topics'] = df.apply(lambda x: fh.flatten_topics(x['topics_universal'], parent_only=True), axis=1)
+    df['flattened_topics_tfidf'] = df.apply(lambda x: pt.flatten_topics(x['topics_universal_tfidf'], parent_only=True), axis=1)
+    df['flattened_topics'] = df.apply(lambda x: pt.flatten_topics(x['topics_universal'], parent_only=True), axis=1)
     df['flattened_speakers'] = df['focal_speakers'].apply(lambda x: ', '.join(x))
     df['flattened_locations'] = df['focal_locations'].apply(lambda x: ', '.join(x))
     df['size'] = 1
 
     if hilite == 'topics_universal':
-        df['hilite'] = df.apply(lambda x: fh.flatten_topics(x['topics_universal'], parent_only=True, max_rank=1), axis=1)
+        df['hilite'] = df.apply(lambda x: pt.flatten_topics(x['topics_universal'], parent_only=True, max_rank=1), axis=1)
         legend_title = 'Primary genre'
     elif hilite == 'topics_universal_tfidf':
-        df['hilite'] = df.apply(lambda x: fh.flatten_topics(x['topics_universal_tfidf'], parent_only=True, max_rank=1), axis=1)
+        df['hilite'] = df.apply(lambda x: pt.flatten_topics(x['topics_universal_tfidf'], parent_only=True, max_rank=1), axis=1)
         legend_title = 'Primary genre'
     elif hilite == 'focal_speakers':
-        df['hilite'] = df.apply(lambda x: fh.flatten_df_list_column(x['focal_speakers'], hilite_color_map, truncate_at=1), axis=1)
+        df['hilite'] = df.apply(lambda x: pt.flatten_df_list_column(x['focal_speakers'], hilite_color_map, truncate_at=1), axis=1)
         legend_title = 'Focal characters'
     elif hilite == 'focal_locations':
-        df['hilite'] = df.apply(lambda x: fh.flatten_df_list_column(x['focal_locations'], hilite_color_map, truncate_at=1), axis=1)
+        df['hilite'] = df.apply(lambda x: pt.flatten_df_list_column(x['focal_locations'], hilite_color_map, truncate_at=1), axis=1)
         legend_title = 'Focal location'
     else:
         df['hilite'] = ''
@@ -113,22 +114,22 @@ def build_speaker_topic_scatter(show_key: str, df: pd.DataFrame, topic_type: str
 
     if not speaker_color_map:
         speakers = df['speaker'].unique()
-        speaker_color_map = fh.generate_speaker_color_discrete_map(show_key, speakers)
+        speaker_color_map = cp.generate_speaker_color_discrete_map(show_key, speakers)
 
     shapes = []
 
     if topic_type == 'mbti':
-        topic_types = fm.mbti_types
-        df['ep_x'] = df['topic_key'].apply(fh.to_mbti_x)
-        df['ep_y'] = df['topic_key'].apply(fh.to_mbti_y)
+        topic_types = stsm.mbti_types
+        df['ep_x'] = df['topic_key'].apply(stsm.to_mbti_x)
+        df['ep_y'] = df['topic_key'].apply(stsm.to_mbti_y)
         title = "Myers-Briggs Temperaments"
         labels = {"ep_x": "←  personal                                            logical  →",
                   "ep_y": "←  present                                            possible  →"}
         high_x = high_y = 4
     elif topic_type == 'dnda':
-        topic_types = fm.dnda_types
-        df['ep_x'] = df['topic_key'].apply(fh.to_dnda_x)
-        df['ep_y'] = df['topic_key'].apply(fh.to_dnda_y)
+        topic_types = stsm.dnda_types
+        df['ep_x'] = df['topic_key'].apply(stsm.to_dnda_x)
+        df['ep_y'] = df['topic_key'].apply(stsm.to_dnda_y)
         title = "D & D Alignments"
         labels = {"ep_x": "←  evil                                                    good  →",
                   "ep_y": "←  chaotic                                               lawful  →"}
@@ -146,16 +147,16 @@ def build_speaker_topic_scatter(show_key: str, df: pd.DataFrame, topic_type: str
         # NOTE Ick. When there's overflow of speakers mapped to a particular topic, things get messy
         # The topic_count and topic_i counters are off by 1, so the checks here are gross
         # Ultimately we df.drop any overflow speaker (current cut-off is 9) mapped to the same topic
-        if topic_count >= len(fm.topic_grid_coord_deltas):
-            topic_count = len(fm.topic_grid_coord_deltas) - 1
+        if topic_count >= len(stsm.topic_grid_coord_deltas):
+            topic_count = len(stsm.topic_grid_coord_deltas) - 1
         topic_i = topics_to_i[topic_key]
         # print(f'topic_key={topic_key} topics_to_counts[topic_key]={topics_to_counts[topic_key]} topic_count={topic_count} topic_i={topic_i}')
         if topic_i >= topic_count:
             # print(f"too many speakers mapped to topic_key={topic_key}, dropping row['speaker']={row['speaker']} from display")
             df.drop(index, inplace=True)
             continue
-        df.at[index, 'ep_x'] = row['ep_x'] + fm.topic_grid_coord_deltas[topic_count][topic_i][0]
-        df.at[index, 'ep_y'] = row['ep_y'] + fm.topic_grid_coord_deltas[topic_count][topic_i][1]
+        df.at[index, 'ep_x'] = row['ep_x'] + stsm.topic_grid_coord_deltas[topic_count][topic_i][0]
+        df.at[index, 'ep_y'] = row['ep_y'] + stsm.topic_grid_coord_deltas[topic_count][topic_i][1]
         topics_to_i[topic_key] += 1
         
     custom_data = ['speaker', 'rank', 'topic_key', 'score']
