+++ app/pages/components.py
@@ -5,6 +5,7 @@ import pandas as pd
 
 import app.fig_builder.fig_helper as fh
 import app.fig_builder.fig_metadata as fm
+import app.figdata_transformer.pandas_transformer as pt
 from app.show_metadata import BGCOLORS_TO_TEXT_COLORS
 from app import utils
 
@@ -172,7 +173,7 @@ def generate_season_episodes_dt(episodes: list) -> dash_table.DataTable:
 
     # field naming and processing
     episodes_df['focal_characters'] = episodes_df['focal_speakers'].apply(lambda x: ', '.join(x))
-    episodes_df['genres'] = episodes_df.apply(lambda x: fh.flatten_topics(x['topics_universal_tfidf'], parent_only=True), axis=1)
+    episodes_df['genres'] = episodes_df.apply(lambda x: pt.flatten_topics(x['topics_universal_tfidf'], parent_only=True), axis=1)
     episodes_df['air_date'] = episodes_df['air_date'].apply(lambda x: x[:10])
     episodes_df.rename(columns={'sequence_in_season': 'episode'}, inplace=True)
