+++ app/pages/episode.py
@@ -14,6 +14,8 @@ import app.fig_builder.plotly_line as pline
 import app.fig_builder.plotly_networkgraph as pgraph
 import app.fig_builder.plotly_scatter as pscat
 import app.fig_builder.plotly_treemap as ptree
+import app.figdata_transformer.color_processor as cp
+import app.figdata_transformer.pandas_transformer as pt
 from app.nlp.nlp_metadata import OPENAI_EMOTIONS
 import app.pages.components as cmp
 from app.show_metadata import ShowKey, TOPIC_COLORS
@@ -493,7 +495,7 @@ def render_episode_search_gantt(show_key: str, episode_key: str, qt: str):
     # build dash datatable
     matching_lines_df.rename(columns={'Task': 'character', 'scene_event': 'line', 'Line': 'dialog'}, inplace=True)
     matching_speakers = list(matching_lines_df['character'].unique())
-    speaker_color_map = fh.generate_speaker_color_discrete_map(show_key, matching_speakers)
+    speaker_color_map = cp.generate_speaker_color_discrete_map(show_key, matching_speakers)
     # TODO matching_lines_df['dialog'] = matching_lines_df['dialog'].apply(convert_markup)
     display_cols = ['character', 'scene', 'line', 'location', 'dialog']
     episode_search_results_dt = cmp.pandas_df_to_dash_dt(matching_lines_df, display_cols, 'character', matching_speakers, speaker_color_map, 
@@ -571,7 +573,7 @@ def render_speaker_3d_network_graph_new(show_key: str, episode_key: str, scale_b
 
     # NOTE where and how to layer in color mapping is a WIP
     speakers = [n['speaker'] for n in speaker_relations_data['nodes']]
-    speaker_colors = fh.generate_speaker_color_discrete_map(show_key, speakers)
+    speaker_colors = cp.generate_speaker_color_discrete_map(show_key, speakers)
     for n in speaker_relations_data['nodes']:
         n['color'] = speaker_colors[n['speaker']].lower() # ugh with the lowercase
 
@@ -607,7 +609,7 @@ def render_speaker_frequency_bar_chart_new(show_key: str, episode_key: str, scal
     df = pd.DataFrame(speakers_for_episode, columns=['speaker', 'agg_score', 'scene_count', 'line_count', 'word_count', 'topics_mbti'])
 
     episode_speaker_names = [s['speaker'] for s in speakers_for_episode]
-    speaker_color_map = fh.generate_speaker_color_discrete_map(show_key, episode_speaker_names)
+    speaker_color_map = cp.generate_speaker_color_discrete_map(show_key, episode_speaker_names)
 
     # TODO incorporate episode-level sentiment into es writer workflow; for now it's a quick lookup in episode-level dfs
     emo_limit = 3
@@ -744,7 +746,7 @@ def render_episode_speaker_topic_scatter(show_key: str, episode_key: str, mbti_c
     episode_speaker_names = [s['speaker'] for s in episode_speakers]
     # indexed_speakers_response = esr.fetch_indexed_speakers(ShowKey(show_key), extra_fields='topics_mbti,topics_dnda', speakers=','.join(episode_speaker_names))
 
-    speaker_color_map = fh.generate_speaker_color_discrete_map(show_key, episode_speaker_names)
+    speaker_color_map = cp.generate_speaker_color_discrete_map(show_key, episode_speaker_names)
 
     # NOTE ended up not using this data downstream
     # mbti_distribution_response = esr.agg_numeric_distrib_into_percentiles(ShowKey(show_key), 'speaker_episode_topics', 'raw_score', constraints='topic_grouping:meyersBriggsKiersey')
@@ -795,7 +797,7 @@ def render_episode_topic_treemap(show_key: str, episode_key: str, ug_score_type:
         r = esr.fetch_episode_topics(ShowKey(show_key), episode_key, tg)
         episode_topics = r['episode_topics']
         df = pd.DataFrame(episode_topics)
-        df = fh.flatten_and_format_topics_df(df, topic_score_types[i])
+        df = pt.flatten_and_format_topics_df(df, topic_score_types[i])
         # build treemap fig
         fig = ptree.build_episode_topic_treemap(df.copy(), tg, topic_score_types[i], max_per_parent=3)
         figs[tg] = fig
