+++ app/pages/series.py
@@ -14,6 +14,8 @@ import app.fig_builder.plotly_bar as pbar
 import app.fig_builder.plotly_gantt as pgantt
 import app.fig_builder.plotly_pie as ppie
 import app.fig_builder.plotly_scatter as pscat
+import app.figdata_transformer.color_processor as cp
+import app.figdata_transformer.pandas_transformer as pt
 import app.nlp.embeddings_factory as ef
 from app.nlp.nlp_metadata import OPENAI_EMOTIONS
 import app.pages.components as cmp
@@ -105,7 +107,7 @@ def layout(show_key: str) -> html.Div:
             universal_genres_parent_topics.append(t['topic_key'])
 
     # series_speaker_names = list(show_metadata[show_key]['regular_cast'].keys()) + list(show_metadata[show_key]['recurring_cast'].keys())
-    speaker_color_map = fh.generate_speaker_color_discrete_map(show_key, all_series_speakers)
+    speaker_color_map = cp.generate_speaker_color_discrete_map(show_key, all_series_speakers)
 
     display_page_end_ts = dt.now()
     display_page_duration = display_page_end_ts - display_page_start_ts
@@ -466,7 +468,7 @@ def render_all_series_episodes_scatter(show_key: str, hilite: str):
         hilite_color_map = TOPIC_COLORS
     elif hilite == 'focal_speakers':
         speakers = list(show_metadata[show_key]['regular_cast'].keys()) + list(show_metadata[show_key]['recurring_cast'].keys())
-        hilite_color_map = fh.generate_speaker_color_discrete_map(show_key, speakers)
+        hilite_color_map = cp.generate_speaker_color_discrete_map(show_key, speakers)
     elif hilite == 'focal_locations':
         scenes_by_location_response = esr.agg_scenes_by_location(ShowKey(show_key))
         scenes_by_location = scenes_by_location_response['scenes_by_location']
@@ -654,7 +656,7 @@ def render_series_search_gantt(show_key: str, series_dialog_qt: str, qt_submit: 
         # build dash datatable
         matching_lines_df.rename(columns={'Task': 'character', 'sequence_in_season': 'episode'}, inplace=True)
         matching_speakers = matching_lines_df['character'].unique()
-        speaker_color_map = fh.generate_speaker_color_discrete_map(show_key, matching_speakers)
+        speaker_color_map = cp.generate_speaker_color_discrete_map(show_key, matching_speakers)
         # TODO matching_lines_df['dialog'] = matching_lines_df['dialog'].apply(convert_markup)
         display_cols = ['episode_key', 'episode_title', 'count', 'season', 'episode', 'info', 'matching_line_count', 'matching_lines']
         episode_search_results_dt = cmp.pandas_df_to_dash_dt(matching_lines_df, display_cols, 'episode_key', matching_speakers, speaker_color_map, 
@@ -753,7 +755,7 @@ def render_series_speaker_listing_dt(show_key: str):
     speakers_df['actor(s)'].fillna('', inplace=True)
     speakers_df['actor(s)'] = speakers_df['actor(s)'].apply(lambda x: ', '.join(x))
 
-    speaker_colors = fh.generate_speaker_color_discrete_map(show_key, speaker_names)
+    speaker_colors = cp.generate_speaker_color_discrete_map(show_key, speaker_names)
 
     speaker_listing_dt = cmp.pandas_df_to_dash_dt(speakers_df, display_cols, 'character', speaker_names, speaker_colors,
                                                   numeric_precision_overrides={'seasons': 0, 'episodes': 0, 'scenes': 0, 'lines': 0, 'words': 0})
@@ -794,7 +796,7 @@ def render_series_speaker_topic_scatter(show_key: str, mbti_count: int, dnda_cou
     indexed_speakers = indexed_speakers_response['speakers']
     # indexed_speakers = fh.flatten_speaker_topics(indexed_speakers, 'mbti', limit_per_speaker=3) 
 
-    speaker_color_map = fh.generate_speaker_color_discrete_map(show_key, series_speaker_names)
+    speaker_color_map = cp.generate_speaker_color_discrete_map(show_key, series_speaker_names)
 
     # flatten episode speaker topic data for each episode speaker
     exploded_speakers_mbti = fh.explode_speaker_topics(indexed_speakers, 'mbti', limit_per_speaker=mbti_count)
@@ -834,7 +836,7 @@ def render_series_topic_figs(show_key: str, topic_grouping: str, score_type: str
         episode_topics_response = esr.fetch_episode_topics(ShowKey(show_key), episode['episode_key'], topic_grouping)
         episode_topic_lists.append(episode_topics_response['episode_topics'])
 
-    series_topics_df, series_parent_topics_df = fh.generate_topic_aggs_from_episode_topics(episode_topic_lists, max_rank=20, max_parent_repeats=2)
+    series_topics_df, series_parent_topics_df = pt.generate_topic_aggs_from_episode_topics(episode_topic_lists, max_rank=20, max_parent_repeats=2)
     ##### TODO end optimization block 
 
     series_topics_pie = ppie.build_topic_aggs_pie(series_topics_df, topic_grouping, score_type)
